requirements.txt
astroid==1.4.8
backports.functools-lru-cache==1.3
backports.shutil-get-terminal-size==1.0.0
configparser==3.5.0
decorator==4.0.10
defusedxml==0.4.1
Django==1.10.2
django-appconf==1.0.2
django-autocomplete-light==3.2.0
django-bootstrap3==7.1.0
django-classy-tags==0.8.0
django-compressor==2.1
django-datetime-widget==0.9.3
django-debug-toolbar==1.6
django-extensions==1.7.4
django-sekizai==0.10.0
enum34==1.1.6
ipython==5.1.0
ipython-genutils==0.1.0
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.5.2
pathlib2==2.1.0
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
ptyprocess==0.5.1
Pygments==2.1.3
pylint==1.6.4
pylint-django==0.7.2
pylint-plugin-utils==0.2.4
python-dateutil==2.5.3
python3-openid==3.0.10
pytz==2016.7
rcssmin==1.0.6
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.2
traitlets==4.3.1
wcwidth==0.1.7
wrapt==1.10.8
